requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# LLM integration
openai>=1.0.0
anthropic>=0.18.0
google-genai>=1.0.0
tiktoken>=0.5.0

# Experiment tracking
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data handling
pydantic>=2.0.0
python-dotenv>=1.0.0

# Utilities
rich>=13.0.0
